atividade8/requirements.txt
openai
gradio
openai-whisper
python-dotenv